_archive/backend_legacy/requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.29.0
google-generativeai==0.3.0
python-dotenv==1.0.0
pydantic>=2.7.0
pydantic-settings==2.1.0
pytest==7.4.0
httpx==0.24.1
pyswisseph==2.10.3.2
geopy==2.4.1
timezonefinder==6.5.7
pytz==2023.3.post1
firebase-admin==6.2.0
email-validator==2.1.0
elevenlabs==1.50.0
stripe
matplotlib
Pillow
